extra checks

+++ site/rest_device.py
@@ -56,10 +56,10 @@ def update_info(aDict):
  tbl_id = { 'devices':'id', 'rackinfo':'device_id' }
  db = GL.DB()
  db.connect()
- if   racked == '0' and aDict.get('devices_rack_id') != 'NULL':
+ if racked and racked == '0' and aDict.get('devices_rack_id') != 'NULL':
   db.do("INSERT INTO rackinfo SET device_id = {} ON DUPLICATE KEY UPDATE rack_unit = 0, rack_size = 1".format(id))
   db.commit()
- elif racked == '1' and aDict.get('devices_rack_id') == 'NULL':
+ elif racked and racked == '1' and aDict.get('devices_rack_id') == 'NULL':
   db.do("DELETE FROM rackinfo WHERE device_id = {}".format(id))
   db.commit()
